requirements.txt
altgraph==0.17
asciimatics==1.11.0
certifi==2020.4.5.1
chardet==3.0.4
future==0.18.2
idna==2.9
mutagen==1.44.0
Pillow==7.1.2
pyfiglet==0.8.post1
PyInstaller==3.6
pytube3==9.6.4
requests==2.23.0
typing-extensions==3.7.4.2
urllib3==1.25.9
wcwidth==0.1.9
